backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
stripe==7.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0

# AI Providers
openai==1.3.0
anthropic==0.8.0

# Job Scraping
aiohttp==3.9.0
beautifulsoup4==4.12.2
apify-client==1.6.0

# HTTP client for APIs
httpx==0.25.2

# Validation
pydantic-settings==2.1.0
